jobportalpr/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.2
django-admin-volt==1.0.11
django-celery-beat==2.7.0
django-filter==25.1
django-modeltranslation==0.19.10
django-timezone-field==7.1
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.5.0
Markdown==3.7
pillow==11.0.0
prompt_toolkit==3.0.50
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.1
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
